webapps2025/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
Django==5.1.7
django-axes==7.0.2
django-crispy-forms==2.3
django-recaptcha==3.0.0
djangorestframework==3.15.2
idna==3.10
psycopg==3.2.6
psycopg-binary==3.2.6
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
